avoid abort due to select() error on systems with sizeof(time_t) < sizeof(long), e.g. Tru64 Unix on Alpha: drop default timeout down to INT_MAX (like in MAIN)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-2-patches@13507 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/snmpd.c
@@ -1083,10 +1083,12 @@ receive(void)
         }
 
         /*
-         * default to sleeping for a really long time
+         * default to sleeping for a really long time. INT_MAX
+         * should be sufficient (eg we don't care if time_t is
+         * a long that's bigger than an int).
          */
         tvp = &timeout;
-        tvp->tv_sec = LONG_MAX;
+        tvp->tv_sec = INT_MAX;
         tvp->tv_usec = 0;
 
         numfds = 0;
